+++ R/venndir-legend.R
@@ -13,7 +13,7 @@
 #' In future, the `setlist` could be derived from the `venndir()` output
 #' object directly. That step likely needs a new function.
 #' 
-#' When using arguments `style="grid"` and `draw_legend=FALSE` the
+#' When using arguments `legend_style="grid"` and `draw_legend=FALSE` the
 #' `grid` `grob` object is returned, so that it can be manipulated
 #' as needed before rendering. Note that in this case, the viewport
 #' will have already been defined and stored into `legend_grob$vp`
@@ -28,24 +28,46 @@
 #' 
 #' @family venndir utility
 #' 
-#' @param venndir_output `Venndir` object returned by `venndir()`, which is used
-#'    to generate the legend counts. When supplied, the `set_colors` are
-#'    also defined by this object.
+#' @param venndir_output `Venndir` object returned by `venndir()`,
+#'    which is used to generate the legend counts.
+#'    When supplied, the `set_colors` are also defined by this object.
 #' @param x `character` string indicating the position of the legend,
-#'    as passed to `graphics::legend()` when `style="base"`.
+#'    as passed to `graphics::legend()` when `legend_style="base"`.
+#'    Not relevant when `legend_style="data.frame"`.
 #' @param set_colors `character` optional vector of R colors, whose names
 #'    should match `names(setlist)`. When not supplied, colors are inferred
 #'    from `venndir_output`, and when that is not supplied, colors are
 #'    defined using `colorjam::rainbowJam()`.
-#' @param setlist `list` DEPRECATED and no longer necessary when
-#'    `venndir_output` is supplied as a `Venndir` object, which already
-#'    contains the `setlist`.
+#' @param unicode `logical` default TRUE, whether to use unicode arrows
+#'    for signed counts, used when `legend_signed=TRUE`.
+#'    This argument is passed to `curate_venn_labels()`.
+#' @param setlist `list` alternative when `venndir_output` is not supplied
+#'    as a `Venndir` object. Used internally by `textvenn()`.
 #' @param keep_newlines `logical` indicating whether to keep newlines
 #'    (line breaks) in the set labels used in the Venn legend.
+#' @param legend_total `logical` default TRUE, whether to include the
+#'    total unique items in the legend.
+#'    When `FALSE` it also sets `legend_percentage=FALSE`.
+#' @param legend_percentage `logical` default NULL, whether to include
+#'    the percentage of total items represented in each set.
+#'    This option is only used when `legend_total=TRUE`.
+#'    * When set to `NULL` it will display percentage only when present
+#'    as a text label, as when `show_labels` include `"p"` or `"P"`.
+#' @param legend_signed `logical` default FALSE, whether to include
+#'    signed counts in the legend.
+#'    * When set to `NULL` it displays signed counts only when available,
+#'    which requires: `overlap_type` is not "overlap", and
+#'    `legend_df` column "type" has entries with "sign".
+#' @param combine_signed `logical` default TRUE, whether to combine multiple
+#'    signed counts into one column (TRUE), or keep separate columns
+#'    (FALSE). Somewhat experimental.
+#'    This option is only used when signed counts are included.
+#' @param combine_size `logical` default TRUE, whether to combine
+#'    counts in the "Size" column with signed counts.
+#'    When `TRUE`, this option also left-aligns the "Size" column.
 #' @param box.lwd `numeric` used to define the box line width,
-#'    as passed to `graphics::legend()` when `style="base"`.
-#' @param style `character` string indicating the style of legend:
-#'    * `"base"` uses `graphics::legend()`
+#'    as passed to `graphics::legend()` when `legend_style="base"`.
+#' @param legend_style `character` string indicating the style of legend:
 #'    * `"grid"` uses `gridExtra::grid.table()`
 #'    * `"data.frame"` only returns the venndir legend `data.frame`.
 #' @param item_type `character` string used as a suffix to the counts
@@ -61,7 +83,7 @@
 #'    is also applied to all cells in the table. Note that the line is
 #'    drawn with line width `lwd`.
 #' @param lwd `numeric` value for line width used for the cell border
-#'    in the legend table, when `style="grid"`.
+#'    in the legend table, when `legend_style="grid"`.
 #' @param x_inset,y_inset `grid::unit` object that defines the inset
 #'    distance away from each plot edge, along the x-axis and y-axis,
 #'    respectively. For example `x_inset=grid::unit(2, "lines")` will
@@ -72,7 +94,7 @@
 #'    than using `"npc"` in terms of fraction of overall plot width
 #'    or height.
 #' @param font_cex `numeric` adjustment to default font sizes. The default
-#'    font size with `style="grid"` uses a 12 point font, so to adjust
+#'    font size with `legend_style="grid"` uses a 12 point font, so to adjust
 #'    to a specific font size like 8 points, use: `font_cex=8/12`
 #' @param table_theme `list` of theme parameters as described in
 #'    `gridExtra::tableGrob()`, and `gridExtra::ttheme_default()`.
@@ -85,10 +107,10 @@
 #'    * `padding` - `grid::unit()` for horizontal and vertical padding
 #'    within each cell
 #' @param draw_legend `logical` indicating whether to draw the resulting
-#'    legend. Because this function is in development, the `style="grid"`
+#'    legend. Because this function is in development, the `legend_style="grid"`
 #'    seems to be preferred, however it may be useful to return the
 #'    `grob` object for manipulation before rendering inside the active
-#'    plot. In that case, use arguments: `style="grid", draw_legend=FALSE`.
+#'    plot. In that case, use arguments: `legend_style="grid", draw_legend=FALSE`.
 #' @param alias `character` vector with optional set aliases, experimental
 #'    feature to allow referring to entries in `setlist` by shorter aliases.
 #'    Note that `names(alias)` must also match `names(setlist)`.
@@ -107,10 +129,12 @@
 #' @param set_suffix `character` string (default `""`) used as optional
 #'    suffix, and previously used `":"` but was changed to `""`.
 #' @param verbose `logical` indicating whether to print verbose output.
-#' @param ... additional arguments are passed to subsequent functions.
+#' @param ... additional arguments are passed to internal functions.
+#'    Notably `curate_venn_labels()` may accept the `curate_df` argument,
+#'    to allow custom labeling.
 #' 
 #' @return `data.frame` with legend information is returned invisibly,
-#'    unless using `style="grid", draw_legend=FALSE` in which case the
+#'    unless using `legend_style="grid", draw_legend=FALSE` in which case the
 #'    legend `grob` object is returned which can then be manipulated
 #'    before rendering.
 #' 
@@ -190,10 +214,16 @@ venndir_legender <- function
  x="bottomleft", 
  setlist=NULL,
  set_colors=NULL,
+ unicode=TRUE,
+ sign_count_delim=NULL,
  keep_newlines=FALSE,
- include_total=TRUE,
+ legend_total=TRUE,
+ legend_percentage=NULL,
+ legend_signed=FALSE,
+ combine_signed=TRUE,
+ combine_size=TRUE,
  box.lwd=0, 
- style=c("grid",
+ legend_style=c("grid",
     "data.frame"),
  item_type="",
  header_color="#000000",
@@ -215,14 +245,59 @@ venndir_legender <- function
  verbose=FALSE,
  ...)
 {
-   style <- match.arg(style);
+   legend_style <- match.arg(legend_style);
 
-   if (length(setlist) == 0) {
-      if (length(venndir_output) > 0 && "Venndir" %in% class(venndir_output)) {
+   if (length(venndir_output) > 0) {
+      if (inherits(venndir_output, "Venndir")) {
          setlist <- venndir_output@setlist;
+      } else {
+         stop("venndir_output must have class 'Venndir'.")
+      }
+   } else if (length(setlist) == 0) {
+      stop("Either setlist or venndir_output must be supplied.")
+   }
+   if (length(sign_count_delim) == 0) {
+      if (length(venndir_output) > 0) {
+         if ("sign_count_delim" %in% names(metadata(venndir_output))) {
+            sign_count_delim <- metadata(venndir_output)$sign_count_delim;
+         }
+      }
+      if (length(sign_count_delim) == 0) {
+         sign_count_delim <- " ";
       }
    }
+   
+   # detect legend_signed
+   if (length(legend_signed) == 0) {
+      if (length(venndir_output) > 0) {
+         if ("overlap" %in% metadata(venndir_output)$overlap_type) {
+            legend_signed <- FALSE
+         } else if (length(metadata(venndir_output)$overlap_type) == 1 &&
+               "signed" %in% venndir_output@label_df$type) {
+            legend_signed <- TRUE;
+         } else {
+            legend_signed <- FALSE;
+         }
+      }
+   }
+   # turn off legend_signed when there are no signed counts
+   if (TRUE %in% legend_signed && length(venndir_output) > 0 &&
+         !"signed" %in% venndir_output@label_df$type) {
+      legend_signed <- FALSE
+   }
+   
+   # detect legend_percentage
+   if (length(legend_percentage) == 0) {
+      legend_percentage <- FALSE;
+      # crudely detect whether percentages are included as labels
+      if (length(venndir_output) > 0 &&
+         any(grepl("[%][* ]*$", venndir_output@label_df$text))) {
+         legend_percentage <- TRUE;
+      }
+   }
+   
    # make sure setlist matches Venn displayed sets
+   legend_labels <- NULL;
    if (length(venndir_output) > 0) {
       use_setlist_names <- intersect(names(setlist),
          venndir_output@label_df$overlap_set)
@@ -230,27 +305,24 @@ venndir_legender <- function
          stop("label_df$overlap_set does not match names(setlist)");
       }
       setlist <- setlist[use_setlist_names];
-   }
-   legend_labels <- NULL;
-   if (length(venndir_output) > 0 &&
-         "legend_label" %in% colnames(venndir_output@jps@polygons)) {
-      matchset <- match(names(setlist),
-         venndir_output@jps@polygons$venn_name)
-      if (any(is.na(matchset))) {
-         stop("Venndir 'venn_name' does not match names(setlist).");
-         # Todo: recover by ignoring legend_labels
-      } else {
-         legend_labels <- jamba::nameVector(
-            venndir_output@jps@polygons$legend_label[matchset],
-            names(setlist));
+      if ("legend_label" %in% colnames(venndir_output@jps@polygons)) {
+         matchset <- match(names(setlist),
+            venndir_output@jps@polygons$venn_name)
+         if (any(is.na(matchset))) {
+            stop("Venndir 'venn_name' does not match names(setlist).");
+            # Todo: recover by ignoring legend_labels
+         } else {
+            legend_labels <- jamba::nameVector(
+               venndir_output@jps@polygons$legend_label[matchset],
+               names(setlist));
+         }
       }
    }
    if (length(legend_labels) == 0) {
       legend_labels <- jamba::nameVector(names(setlist));
    }
-   # jamba::printDebug("legend_labels:");print(legend_labels);# debug
-   
-   # custom function to replace newline style with \n
+
+   # custom function to replace newline legend_style with \n
    fix_setlist_names <- function
    (setlist,
     use_newline="\n",
@@ -277,18 +349,20 @@ venndir_legender <- function
       if (length(venndir_output) == 0) {
          stop("Either setlist or venndir_output must be provided.");
       }
-      vodf <- venndir_output@jps@polygons;
-      vodf1 <- subset(vodf, type %in% "overlap");
-      total_items <- sum(vodf1$venn_counts);
-      # vodf1[, c("venn_name", "venn_counts")]
-      vodf_names <- strsplit(vodf1$venn_name, "&");
-      vodf_levels <- unique(unlist(vodf_names));
-      vodf_counts <- rep(vodf1$venn_counts, lengths(vodf_names));
-      vodf_split <- factor(unlist(vodf_names), levels=vodf_levels);
-      setlist_sizes <- sapply(split(vodf_counts, vodf_split), sum);
-      setlist <- lapply(setlist_sizes, function(i){
-         rep(1, length.out=i)
-      })
+      setlist <- setlist(venndir_output);
+      ## old logic
+      # vodf <- venndir_output@jps@polygons;
+      # vodf1 <- subset(vodf, type %in% "overlap");
+      # total_items <- sum(vodf1$venn_counts);
+      # # vodf1[, c("venn_name", "venn_counts")]
+      # vodf_names <- strsplit(vodf1$venn_name, "&");
+      # vodf_levels <- unique(unlist(vodf_names));
+      # vodf_counts <- rep(vodf1$venn_counts, lengths(vodf_names));
+      # vodf_split <- factor(unlist(vodf_names), levels=vodf_levels);
+      # setlist_sizes <- sapply(split(vodf_counts, vodf_split), sum);
+      # setlist <- lapply(setlist_sizes, function(i){
+      #    rep(1, length.out=i)
+      # })
    }
    
    # pick out venn colors if available
@@ -360,7 +434,7 @@ venndir_legender <- function
    }
    
    # optionally include total unique items
-   if (TRUE %in% include_total) {
+   if (TRUE %in% legend_total) {
       if (length(venndir_output) > 0) {
          # venndir_legender(setlist=setlist, venndir_output=vo, x="bottomright")
          total_items <- sum(vodf_all$venn_counts, na.rm=TRUE);
@@ -418,6 +492,78 @@ venndir_legender <- function
       set=names(setlist),
       size=lengths(setlist))
 
+   # optional signed counts
+   if (TRUE %in% legend_signed) {
+      sv <- jamba::rbindList(lapply(seq_along(setlist), function(isetlist){
+         idf <- signed_overlaps(setlist[isetlist],
+            overlap_type="each",
+            return_items=FALSE);
+         idf[, setdiff(colnames(idf), names(setlist[isetlist])), drop=FALSE]
+      }))
+      # total_items <- sum(sv$count, na.rm=TRUE);
+      gCounts <- lapply(jamba::nameVector(unique(sv$sets)), function(i){
+         j <- subset(sv, sets %in% i);
+         jamba::nameVector(j$count, j$overlap_label)
+      });
+      
+      gbase_labels <- curate_venn_labels(
+         x=names(unlist(unname(gCounts))),
+         unicode=unicode,
+         type="sign",
+         ...)
+      gbase_colors <- curate_venn_labels(
+         x=names(unlist(unname(gCounts))),
+         unicode=unicode,
+         type="color",
+         ...)
+         # curate_df=curate_df);
+      big.mark <- ",";
+      use_sign_count_delim <- sub(" ", "", sign_count_delim);
+      gcount_labels <- sapply(seq_along(unlist(gCounts)), function(i){
+         ilabel <- paste0(
+            gbase_labels[i],
+            use_sign_count_delim,
+            format(trim=TRUE,
+               big.mark=big.mark,
+               unlist(gCounts)[i]));
+      });
+      gcount_df <- data.frame(check.names=FALSE,
+         set_name=rep(names(gCounts), lengths(gCounts)),
+         count=unlist(unname(gCounts)),
+         label=gbase_labels,
+         count_label=gcount_labels,
+         count_label_color=gbase_colors)
+      gcdf <- data.frame(check.names=FALSE,
+         jamba::rbindList(lapply(split(gcount_df, gcount_df$set_name), function(idf){
+            unname(unlist(
+               strsplit(
+                  jamba::pasteByRow(
+                     idf[, c("count_label_color", "count_label"), drop=FALSE],
+                     sep="!!"), "!!")))
+         })))
+      gcdf1 <- gcdf[, seq(from=1, to=ncol(gcdf), by=2), drop=FALSE];
+      # each count in its own column, mostly so it can be colorized
+      gcdf2 <- gcdf[, seq(from=2, to=ncol(gcdf), by=2), drop=FALSE];
+      # replace with "(# up, # dn)" format
+      gcolnames1 <- c(
+         # " (Sign)",
+         " Sign",
+         sapply(seq_len(ncol(gcdf2)), function(k1){
+            paste(rep(" ", k1), collapse="")}));
+      gcolnames <- tail(gcolnames1, -1)
+      colnames(gcdf2) <- gcolnames
+      colnames(gcdf1) <- colnames(gcdf2);
+      if (TRUE %in% combine_signed) {
+         gcdf2 <- data.frame(check.names=FALSE,
+            ` Sign`=paste0(" (", jamba::pasteByRow(gcdf2, sep=", "), ")"))
+         # colnames(gcdf2) <- gcolnames[1]
+      } else if (TRUE %in% combine_size) {
+         colnames(gcdf2) <- jamba::makeNames(rep("Sign", ncol(gcdf2)));
+      }
+      rownames(gcdf2) <- rownames(gcdf1);
+      # print(gcdf2);
+   }
+   
    # optional labels
    if (length(labels) > 0) {
       if (!all(names(setlist) %in% names(labels))) {
@@ -455,7 +601,7 @@ venndir_legender <- function
    }
    
    # optionally include total item count
-   if (TRUE %in% include_total) {
+   if (TRUE %in% legend_total) {
       gridlegend_df2 <- data.frame(check.names=FALSE,
          stringsAsFactors=FALSE,
          set=paste0("Total", ifelse(nchar(item_type) > 0, " ", ""),
@@ -474,8 +620,43 @@ venndir_legender <- function
       legend_df <- rbind(legend_df, legend_df2);
    }
 
+   # optionally add percentage
+   if (TRUE %in% legend_percentage && TRUE %in% legend_total) {
+      pct_total <- round(
+         1000 * gridlegend_df$size / tail(gridlegend_df$size, 1)) / 10;
+      gcdf2p <- data.frame(check.names=FALSE,
+         `Percentage`=paste(format(pct_total, digits=3, trim=TRUE), "%"))
+      rownames(gcdf2p) <- gridlegend_df$set;
+      if (TRUE %in% legend_signed) {
+         matchgc <- match(rownames(gcdf2p), rownames(gcdf2))
+         gcdf2 <- data.frame(check.names=FALSE,
+            gcdf2[matchgc, , drop=FALSE],
+            gcdf2p)
+         rownames(gcdf2) <- rownames(gcdf2p);
+      } else {
+         gcdf2 <- gcdf2p;
+      }
+   }
+   
+   # add signed/percentage columns to legend
+   if (TRUE %in% legend_signed || TRUE %in% legend_percentage) {
+      matchgc <- match(gridlegend_df$set, rownames(gcdf2));
+      add_df <- gcdf2[match(gridlegend_df$set, rownames(gcdf2)), , drop=FALSE];
+      add_df[is.na(add_df)] <- ""
+      addc_df <- gcdf1[match(gridlegend_df$set, rownames(gcdf2)), , drop=FALSE];
+      addc_df[is.na(addc_df)] <- ""
+      gridlegend_df[, colnames(add_df)] <- add_df
+      # gridlegend_df
+      # legend_df
+   }
+   
+   # # return data.frame
+   # if ("data.frame" %in% legend_style) {
+   #    return(list(gridlegend_df=gridlegend_df, legend_df=legend_df))
+   # }
+   
    # render legend
-   if ("grid" %in% style) {
+   if (any(c("grid", "data.frame") %in% legend_style)) {
       # prepare graphical object
       gridlegend_df_use <- gridlegend_df;
       gridlegend_df_use$set <- paste0(gridlegend_df$set, set_suffix);
@@ -490,6 +671,31 @@ venndir_legender <- function
       }
       if (length(header_color) == 0 || nchar(header_color) == 0) {
          header_color <- "#00000000"
+      }
+      
+      ## optionally combine Size with Sign
+      # - Note: Include percentage made the final string confusing,
+      #   since percentage is not in parentheses
+      # if (TRUE %in% legend_signed || TRUE %in% legend_percentage) {
+      if (TRUE %in% combine_size && TRUE %in% legend_signed) {
+         signcol <- grep("Sign", value=TRUE, colnames(gridlegend_df_use));
+         # signcol <- grep("Sign|Percentage", value=TRUE,
+         #    colnames(gridlegend_df_use));
+         gridlegend_df_use$size <- sub("  [(]", " (",
+            jamba::pasteByRow(
+               gridlegend_df_use[, c("size", signcol), drop=FALSE], sep=" "));
+         keepcols <- setdiff(colnames(gridlegend_df_use), signcol)
+         gridlegend_df_use <- gridlegend_df_use[, keepcols, drop=FALSE];
+         gridlegend_df_use <- jamba::renameColumn(gridlegend_df_use,
+            from="size", to="Size ")
+      }
+      
+      if ("data.frame" %in% legend_style) {
+         colnames(gridlegend_df_use) <- jamba::ucfirst(
+            colnames(gridlegend_df_use));
+         rownames(gridlegend_df_use) <- seq_len(
+            nrow(gridlegend_df_use));
+         return(gridlegend_df_use)
       }
       
       ## table theme
@@ -519,8 +725,14 @@ venndir_legender <- function
          # align_list <- lapply(jamba::nameVectorN(gridlegend_df_use), function(icol){
          #    list(fg_params=list(hjust=0))
          # })
-         use_hjust1 <- rep(c(0, 1),
-            c(ncol(gridlegend_df_use) - 1, 1))
+         # use_hjust1 <- ifelse(
+         #    grepl("Set|Label|Sign", colnames(gridlegend_df_use)),
+         #    0, 1)
+         use_hjust1 <- ifelse(
+            grepl("Size$|Percent", colnames(gridlegend_df_use)),
+            1, 0)
+         # use_hjust1 <- rep(c(0, 1),
+         #    c(ncol(gridlegend_df_use) - 1, 1))
          use_hjust <- rep(use_hjust1,
             each=nrow(gridlegend_df_use))
          if (length(legend_padding) == 0) {